Migrate FormatTime tests to munit
